Moved settings logic to separate file
